moved app component to components folder too
